fix preload images
