+++ packages/amplication-client/src/Resource/create-resource/wizard-pages/CreateServiceAuth.tsx
@@ -6,9 +6,9 @@ import { WizardStepProps } from "./interfaces";
 import authModuleImage from "../../../assets/images/auth-module.svg";
 import ImgSvg from "./ImgSvg";
 
-const AuthCoreSvg = ImgSvg({ image: authModuleImage });
-
 const CreateServiceAuth: React.FC<WizardStepProps> = ({ formik }) => {
+  const AuthCoreSvg = ImgSvg({ image: authModuleImage });
+
   const handleDatabaseSelect = useCallback(
     (database: string) => {
       formik.setValues(
